+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 ## 0.10 (2021-xx-xx)
 
+* Implement [Supervisor Issue #177](https://github.com/Supervisor/supervisor/issues/177).
+  It is possible to update dynamically the program numprocs using the new ``update_numprocs`` XML-RPC. 
+
 * Add targets **Python 3.7** and **Python 3.8**.
 
 * Update documentation.